Fixes game mode on Huntsman V2 while maintaining functionality for other keyboards such as the Blackwidow Chroma V2
- Added special key related device files to the mouse protocol interface for Huntsman V2
- Modified set_game_mode to handle devices files on either interface
- Tests keyboards that use both variations

+++ daemon/openrazer_daemon/dbus_services/dbus_methods/chroma_keyboard.py
@@ -76,11 +76,14 @@ def set_game_mode(self, enable):
 
     driver_path = self.get_driver_path('game_led_state')
 
-    for kb_int in self.additional_interfaces:
-        super_file = os.path.join(kb_int, 'key_super')
-        alt_tab = os.path.join(kb_int, 'key_alt_tab')
-        alt_f4 = os.path.join(kb_int, 'key_alt_f4')
+    # Fixes setting Game Mode/LED for Huntsman V2 w/ device
+    # files added to the usual MOUSE PROTOCOL Interface.
+    # Other devices seem to use the KEYBOARD PROTOCOL interface.
+    super_file = self.get_driver_path('key_super')
+    alt_tab = self.get_driver_path('key_alt_tab')
+    alt_f4 = self.get_driver_path('key_alt_f4')
 
+    if os.path.exists(super_file):
         if enable:
             open(super_file, 'wb').write(b'\x01')
             open(alt_tab, 'wb').write(b'\x01')
@@ -89,6 +92,20 @@ def set_game_mode(self, enable):
             open(super_file, 'wb').write(b'\x00')
             open(alt_tab, 'wb').write(b'\x00')
             open(alt_f4, 'wb').write(b'\x00')
+    else:
+        for kb_int in self.additional_interfaces:
+            super_file = os.path.join(kb_int, 'key_super')
+            alt_tab = os.path.join(kb_int, 'key_alt_tab')
+            alt_f4 = os.path.join(kb_int, 'key_alt_f4')
+            
+            if enable:
+                open(super_file, 'wb').write(b'\x01')
+                open(alt_tab, 'wb').write(b'\x01')
+                open(alt_f4, 'wb').write(b'\x01')
+            else:
+                open(super_file, 'wb').write(b'\x00')
+                open(alt_tab, 'wb').write(b'\x00')
+                open(alt_f4, 'wb').write(b'\x00')
 
     with open(driver_path, 'w') as driver_file:
         if enable:
